+++ zcash_primitives/CHANGELOG.md
@@ -20,7 +20,7 @@ workspace.
 - MSRV is now 1.85.1.
 - Migrated to `orchard 0.12`, `sapling-crypto 0.6`.
 - `zcash_primitives::transaction::builder`:
-  - `Error` has new `Coinbase` and `ProposedVersionIncompatible` variants.
+  - `Error` has new `Coinbase` and `TargetIncompatible` variants.
   - `Builder::add_orchard_output`'s `value` parameter now has type `Zatoshis`
     instead of `u64`.
   - `Builder::add_transparent_input` now takes a `zcash_transparent::builder::TransparentInputInfo`